parsed fixed, changed duration from uitn32_t to double since now we parse pcaps differently

+++ contrib/utils/model/swift-utils.h
@@ -28,7 +28,7 @@ namespace ns3 {
     struct flow_metadata_new {
         std::string prefix;
         uint32_t packets;
-        uint32_t duration;
+        double duration;
         uint64_t bytes;
         double rtt;
     };
